add columns for collection status in alembic
